add rules test file separate from scheduler
