+++ tests/python/pants_test/engine/test_scheduler.py
@@ -15,13 +15,12 @@ from pants.engine.addressable import Addresses
 from pants.engine.engine import LocalSerialEngine
 from pants.engine.nodes import (ConflictingProducersError, DependenciesNode, Return, SelectNode,
                                 Throw, Waiting)
-from pants.engine.rules import NodeBuilder, RulesetValidator
 from pants.engine.scheduler import (CompletedNodeException, IncompleteDependencyException,
                                     ProductGraph)
 from pants.engine.selectors import Select, SelectDependencies, SelectVariant
 from pants.util.contextutil import temporary_dir
 from pants_test.engine.examples.planners import (ApacheThriftJavaConfiguration, Classpath, GenGoal,
-                                                 Goal, Jar, JavaSources, ThriftSources,
+                                                 Jar, JavaSources, ThriftSources,
                                                  setup_json_scheduler)
 
 
@@ -420,82 +419,3 @@ class ProductGraphTest(unittest.TestCase):
             associated_entry.node in before_nodes,
             'node:\n{}\nis still associated with:\n{}\nin {}'.format(node, associated_entry.node, entry)
           )
-
-
-class AGoal(Goal):
-
-  @classmethod
-  def products(cls):
-    return [A]
-
-
-class A(object):
-  pass
-
-
-class B(object):
-  pass
-
-
-def noop(*args):
-  pass
-
-
-class SubA(A):
-  pass
-
-
-class RulesetValidatorTest(unittest.TestCase):
-  def test_ruleset_with_missing_product_type(self):
-    validator = RulesetValidator(NodeBuilder.create([(A, (Select(B),), noop)]),
-      goal_to_product=dict(),
-      root_subject_types=tuple())
-    with self.assertRaises(ValueError):
-      validator.validate()
-
-  def test_ruleset_with_with_selector_only_provided_as_root_subject(self):
-
-    validator = RulesetValidator(NodeBuilder.create([(A, (Select(B),), noop)]),
-      goal_to_product=dict(),
-      root_subject_types=(B,))
-
-    validator.validate()
-
-  def test_ruleset_with_superclass_of_selected_type_produced(self):
-
-    rules = [
-      (A, (Select(B),), noop),
-      (B, (Select(SubA),), noop)
-    ]
-    validator = RulesetValidator(NodeBuilder.create(rules),
-      goal_to_product=dict(),
-      root_subject_types=tuple())
-
-    validator.validate()
-
-  def test_ruleset_with_goal_not_produced(self):
-
-    rules = [
-      (B, (Select(SubA),), noop)
-    ]
-    validator = RulesetValidator(NodeBuilder.create(rules),
-      goal_to_product={'goal-name': AGoal},
-      root_subject_types=tuple())
-    with self.assertRaises(ValueError):
-      validator.validate()
-
-      # products that are not used
-      # selectors of different types that cannot be provided
-      # maybe this needs to be a separate type.
-      # :/
-      #def test_ruleset_with_no_usage_of_product_type(self):
-      #  class A(object):
-      #    pass
-      #  class B(object):
-      #    pass
-      #  def noop(*args):
-      #    pass
-      #
-      #  with self.assertRaises(ValueError):
-      #    RulesetValidator(NodeBuilder.create([(A, (Select(B),), noop)]), None, None)
-      #
